+++ packages/pob-lcov-reporter/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.2.0](https://github.com/christophehurpeau/pob/compare/pob-lcov-reporter@5.1.1...pob-lcov-reporter@5.2.0) (2021-09-25)
+
+
+### Features
+
+* **deps:** update jest monorepo to v27.1.0 ([#1030](https://github.com/christophehurpeau/pob/issues/1030)) ([0ca73d8](https://github.com/christophehurpeau/pob/commit/0ca73d8f4ed0ad238860251dd610480b3f45436d))
+* **deps:** update jest monorepo to v27.1.1 ([#1032](https://github.com/christophehurpeau/pob/issues/1032)) ([2221a09](https://github.com/christophehurpeau/pob/commit/2221a0933c1fe14cb7403f0232fb342785b55861))
+* **deps:** update jest monorepo to v27.2.0 ([#1041](https://github.com/christophehurpeau/pob/issues/1041)) ([c9d4e56](https://github.com/christophehurpeau/pob/commit/c9d4e567f496c8c9c308247210c71be41eb42404))
+* **deps:** update jest monorepo to v27.2.1 ([#1048](https://github.com/christophehurpeau/pob/issues/1048)) ([9014cd5](https://github.com/christophehurpeau/pob/commit/9014cd5865fd0c9ec9de76404c4c78db533e7d33))
+
+
+
+
+
 ## [5.1.1](https://github.com/christophehurpeau/pob/compare/pob-lcov-reporter@5.1.0...pob-lcov-reporter@5.1.1) (2021-09-05)
 
 **Note:** Version bump only for package pob-lcov-reporter